update bot remove unused comments add custom command check

+++ src/lib/commands/rem.py
@@ -7,6 +7,11 @@ def rem(args):
     db = Database()
     command = args[0].lower()
     channel = globals.global_channel
-    db.remove_command(command, channel)
-    return "{0} removed from Ravenbot007's custom commands!".format(
-        command)
+    command_data = db.get_command(command, channel)
+    print channel
+    if command_data:
+        db.remove_command(command, channel)
+        return "{0} removed from Ravenbot007's custom commands!".format(
+            command)
+    else:
+        return "{0} not found.".format(command)
